feat: unify sqlite bootstrap for stale db migration
