+++ packages/core/src/infra/mod.rs
@@ -6,12 +6,127 @@
 //! - `llm.rs` - LLM 客户端 (Claude, OpenAI)
 //! - `schema.sql` - 数据库 Schema
 
+use crate::error::{InfraError, InfraResult};
+use rusqlite::Connection;
+
 mod contract;
 mod db_migration;
 mod llm;
 mod paths;
 pub mod quota_state;
 mod sqlite;
+
+const FTS_BOOTSTRAP_USER_VERSION: i64 = 1;
+const ALLOWED_COLUMN_EXISTS_TABLES: &[&str] = &["items", "documents"];
+
+pub fn prepare_sqlite_db(conn: &Connection) -> InfraResult<()> {
+    conn.execute_batch(include_str!("schema.sql"))
+        .map_err(|e| InfraError::Database(e.to_string()))?;
+
+    migrate_items_add_document_columns(conn)?;
+    migrate_documents_url_unique(conn)?;
+    let _ = maybe_rebuild_fts_index(conn)?;
+
+    Ok(())
+}
+
+pub(crate) fn maybe_rebuild_fts_index(conn: &Connection) -> InfraResult<bool> {
+    let user_version = conn
+        .pragma_query_value(None, "user_version", |row| row.get::<_, i64>(0))
+        .map_err(|e| InfraError::Database(e.to_string()))?;
+    if user_version < FTS_BOOTSTRAP_USER_VERSION {
+        let item_count: i64 = conn
+            .query_row("SELECT COUNT(*) FROM items", [], |row| row.get(0))
+            .map_err(|e| InfraError::Database(e.to_string()))?;
+        let rebuilt = if item_count > 0 {
+            conn.execute("INSERT INTO items_fts(items_fts) VALUES('rebuild')", [])
+                .map_err(|e| InfraError::Database(e.to_string()))?;
+            true
+        } else {
+            false
+        };
+        conn.pragma_update(None, "user_version", FTS_BOOTSTRAP_USER_VERSION)
+            .map_err(|e| InfraError::Database(e.to_string()))?;
+        return Ok(rebuilt);
+    }
+
+    if conn
+        .execute(
+            "INSERT INTO items_fts(items_fts) VALUES('integrity-check')",
+            [],
+        )
+        .is_ok()
+    {
+        return Ok(false);
+    }
+
+    conn.execute("INSERT INTO items_fts(items_fts) VALUES('rebuild')", [])
+        .map_err(|e| InfraError::Database(e.to_string()))?;
+    conn.execute(
+        "INSERT INTO items_fts(items_fts) VALUES('integrity-check')",
+        [],
+    )
+    .map_err(|e| InfraError::Database(e.to_string()))?;
+    Ok(true)
+}
+
+pub(crate) fn column_exists(conn: &Connection, table: &str, column: &str) -> InfraResult<bool> {
+    if !ALLOWED_COLUMN_EXISTS_TABLES.contains(&table) {
+        return Err(InfraError::Database(format!("unknown table: {table}")));
+    }
+    let mut stmt = conn
+        .prepare(&format!("PRAGMA table_info({table})"))
+        .map_err(|e| InfraError::Database(e.to_string()))?;
+    let names: Vec<String> = stmt
+        .query_map([], |row| row.get::<_, String>(1))
+        .map_err(|e| InfraError::Database(e.to_string()))?
+        .filter_map(|r| r.ok())
+        .collect();
+    Ok(names.iter().any(|n| n == column))
+}
+
+fn migrate_items_add_document_columns(conn: &Connection) -> InfraResult<()> {
+    let has_document_id = column_exists(conn, "items", "document_id")?;
+    if !has_document_id {
+        conn.execute_batch("ALTER TABLE items ADD COLUMN document_id TEXT")
+            .map_err(|e| InfraError::Database(e.to_string()))?;
+    }
+    let has_excerpt = column_exists(conn, "items", "excerpt")?;
+    if !has_excerpt {
+        conn.execute_batch("ALTER TABLE items ADD COLUMN excerpt TEXT")
+            .map_err(|e| InfraError::Database(e.to_string()))?;
+    }
+    conn.execute_batch("CREATE INDEX IF NOT EXISTS idx_items_document ON items(document_id)")
+        .map_err(|e| InfraError::Database(e.to_string()))?;
+    Ok(())
+}
+
+fn migrate_documents_url_unique(conn: &Connection) -> InfraResult<()> {
+    conn.execute_batch(
+        "UPDATE items
+         SET document_id = (
+             SELECT d_keep.id
+             FROM documents d_dup
+             JOIN documents d_keep ON d_dup.url = d_keep.url
+             WHERE d_dup.id = items.document_id
+               AND d_keep.rowid = (SELECT MIN(rowid) FROM documents WHERE url = d_dup.url)
+         )
+         WHERE document_id IN (
+             SELECT id FROM documents
+             WHERE rowid NOT IN (SELECT MIN(rowid) FROM documents GROUP BY url)
+         )",
+    )
+    .map_err(|e| InfraError::Database(e.to_string()))?;
+    conn.execute_batch(
+        "DELETE FROM documents WHERE rowid NOT IN (SELECT MIN(rowid) FROM documents GROUP BY url)",
+    )
+    .map_err(|e| InfraError::Database(e.to_string()))?;
+    conn.execute_batch("DROP INDEX IF EXISTS idx_documents_url")
+        .map_err(|e| InfraError::Database(e.to_string()))?;
+    conn.execute_batch("CREATE UNIQUE INDEX IF NOT EXISTS idx_documents_url ON documents(url)")
+        .map_err(|e| InfraError::Database(e.to_string()))?;
+    Ok(())
+}
 
 // 公共 API
 pub use contract::{
