+++ packages/core/src/infra/sqlite/ops.rs
@@ -4,122 +4,21 @@ use crate::knowledge::{Item, ItemType};
 use chrono::{DateTime, Utc};
 use rusqlite::{params, Connection, OptionalExtension};
 
-const FTS_BOOTSTRAP_USER_VERSION: i64 = 1;
+#[cfg(test)]
+const FTS_BOOTSTRAP_USER_VERSION: i64 = super::super::FTS_BOOTSTRAP_USER_VERSION;
 
 pub(super) fn init_schema(conn: &Connection) -> InfraResult<()> {
-    conn.execute_batch(include_str!("../schema.sql"))
-        .map_err(|e| InfraError::Database(e.to_string()))?;
-
-    migrate_items_add_document_columns(conn)?;
-    migrate_documents_url_unique(conn)?;
-    let _ = maybe_rebuild_fts_index(conn)?;
-
-    Ok(())
+    super::super::prepare_sqlite_db(conn)
 }
 
-fn migrate_items_add_document_columns(conn: &Connection) -> InfraResult<()> {
-    let has_document_id = column_exists(conn, "items", "document_id")?;
-    if !has_document_id {
-        conn.execute_batch("ALTER TABLE items ADD COLUMN document_id TEXT")
-            .map_err(|e| InfraError::Database(e.to_string()))?;
-    }
-    let has_excerpt = column_exists(conn, "items", "excerpt")?;
-    if !has_excerpt {
-        conn.execute_batch("ALTER TABLE items ADD COLUMN excerpt TEXT")
-            .map_err(|e| InfraError::Database(e.to_string()))?;
-    }
-    conn.execute_batch("CREATE INDEX IF NOT EXISTS idx_items_document ON items(document_id)")
-        .map_err(|e| InfraError::Database(e.to_string()))?;
-    Ok(())
-}
-
-fn migrate_documents_url_unique(conn: &Connection) -> InfraResult<()> {
-    // Remap items that point to a duplicate document to the kept document (earliest rowid
-    // per URL) before deleting duplicates, so no items become orphans after the migration.
-    conn.execute_batch(
-        "UPDATE items
-         SET document_id = (
-             SELECT d_keep.id
-             FROM documents d_dup
-             JOIN documents d_keep ON d_dup.url = d_keep.url
-             WHERE d_dup.id = items.document_id
-               AND d_keep.rowid = (SELECT MIN(rowid) FROM documents WHERE url = d_dup.url)
-         )
-         WHERE document_id IN (
-             SELECT id FROM documents
-             WHERE rowid NOT IN (SELECT MIN(rowid) FROM documents GROUP BY url)
-         )",
-    )
-    .map_err(|e| InfraError::Database(e.to_string()))?;
-    // Remove duplicate URLs keeping the earliest-inserted row before adding constraint.
-    conn.execute_batch(
-        "DELETE FROM documents WHERE rowid NOT IN (SELECT MIN(rowid) FROM documents GROUP BY url)",
-    )
-    .map_err(|e| InfraError::Database(e.to_string()))?;
-    // Drop old non-unique index (may not exist on fresh databases).
-    conn.execute_batch("DROP INDEX IF EXISTS idx_documents_url")
-        .map_err(|e| InfraError::Database(e.to_string()))?;
-    // Create unique index so concurrent inserts of the same URL fail fast.
-    conn.execute_batch("CREATE UNIQUE INDEX IF NOT EXISTS idx_documents_url ON documents(url)")
-        .map_err(|e| InfraError::Database(e.to_string()))?;
-    Ok(())
-}
-
-const ALLOWED_TABLES: &[&str] = &["items", "documents"];
-
+#[cfg(test)]
 fn column_exists(conn: &Connection, table: &str, column: &str) -> InfraResult<bool> {
-    if !ALLOWED_TABLES.contains(&table) {
-        return Err(InfraError::Database(format!("unknown table: {table}")));
-    }
-    let mut stmt = conn
-        .prepare(&format!("PRAGMA table_info({table})"))
-        .map_err(|e| InfraError::Database(e.to_string()))?;
-    let names: Vec<String> = stmt
-        .query_map([], |row| row.get::<_, String>(1))
-        .map_err(|e| InfraError::Database(e.to_string()))?
-        .filter_map(|r| r.ok())
-        .collect();
-    Ok(names.iter().any(|n| n == column))
+    super::super::column_exists(conn, table, column)
 }
 
+#[cfg(test)]
 fn maybe_rebuild_fts_index(conn: &Connection) -> InfraResult<bool> {
-    let user_version = conn
-        .pragma_query_value(None, "user_version", |row| row.get::<_, i64>(0))
-        .map_err(|e| InfraError::Database(e.to_string()))?;
-    if user_version < FTS_BOOTSTRAP_USER_VERSION {
-        let item_count: i64 = conn
-            .query_row("SELECT COUNT(*) FROM items", [], |row| row.get(0))
-            .map_err(|e| InfraError::Database(e.to_string()))?;
-        let rebuilt = if item_count > 0 {
-            conn.execute("INSERT INTO items_fts(items_fts) VALUES('rebuild')", [])
-                .map_err(|e| InfraError::Database(e.to_string()))?;
-            true
-        } else {
-            false
-        };
-        conn.pragma_update(None, "user_version", FTS_BOOTSTRAP_USER_VERSION)
-            .map_err(|e| InfraError::Database(e.to_string()))?;
-        return Ok(rebuilt);
-    }
-
-    if conn
-        .execute(
-            "INSERT INTO items_fts(items_fts) VALUES('integrity-check')",
-            [],
-        )
-        .is_ok()
-    {
-        return Ok(false);
-    }
-
-    conn.execute("INSERT INTO items_fts(items_fts) VALUES('rebuild')", [])
-        .map_err(|e| InfraError::Database(e.to_string()))?;
-    conn.execute(
-        "INSERT INTO items_fts(items_fts) VALUES('integrity-check')",
-        [],
-    )
-    .map_err(|e| InfraError::Database(e.to_string()))?;
-    Ok(true)
+    super::super::maybe_rebuild_fts_index(conn)
 }
 pub(super) fn find_by_id(conn: &Connection, id: &str) -> InfraResult<Option<Item>> {
     let mut stmt = conn
